refactor: refactoring chat container component
