+++ src/components/ChatHeader/styles.scss
@@ -1,0 +1,71 @@
+@import '../../assets/scss/variables';
+@import '../../assets/scss/mixins';
+
+.chat-container-header {
+  position: sticky;
+  top: 0;
+  @include flex-row-space-between-center;
+  gap: 1rem;
+  height: 5rem;
+  padding: 1rem;
+  &[data-theme="light"] {
+    border-bottom: 1px solid rgba($color-black, 0.24);
+  }
+  &[data-theme="dark"] {
+    border-bottom: 1px solid rgba($color-white, 0.24);
+  }
+
+  > .chat-icon {
+    height: 3rem;
+    width: auto;
+    object-fit: cover;
+    border-radius: 50%;
+    cursor: pointer;
+  }
+
+  > .chat-infos {
+    flex: 1;
+    @include flex-column-center-start;
+
+    > .chat-name {
+      font-size: $font-size-medium;
+      font-weight: $font-weight-semi-bold;
+      color: $color-black;
+
+      &[data-theme="light"] {
+        color: $color-black;
+      }
+      &[data-theme="dark"] {
+        color: $color-white;
+      }
+    }
+
+    > .members {
+      font-size: $font-size-small;
+      font-weight: $font-weight-semi-bold;
+      &[data-theme="light"] {
+        color: $color-neutral-600;
+      }
+      &[data-theme="dark"] {
+        color: $color-neutral-500;
+      }
+    }
+
+  }
+
+  > .chat-options {
+
+    > .options-icon {
+      width: 1.6rem;
+      height: auto;
+      cursor: pointer;
+
+      &[data-theme="light"] {
+        fill: $color-black;
+      }
+      &[data-theme="dark"] {
+        fill: $color-white;
+      }
+    }
+  }
+}
