+++ src/components/Header/styles.scss
@@ -121,8 +121,12 @@
       cursor: pointer;
 
       &:hover {
-        filter: brightness(0.9);
-        transition: filter 200ms;
+        opacity: 0.7;
+        transition: opacity 200ms;
+      }
+
+      &[data-default-img="true"][data-theme="dark"] {
+        filter: invert(1);
       }
     }
