+++ src/components/MessagesContainer/styles.scss
@@ -1,0 +1,93 @@
+@import '../../assets/scss/variables';
+@import '../../assets/scss/mixins';
+
+.messages-container {
+  flex: 1;
+  @include flex-column-start-start;
+  min-height: 70vh;
+  min-height: 70svh;
+  max-height: 70vh;
+  max-height: 70svh;
+  overflow-y: auto;
+  padding: .5rem 1rem 3.3rem 1rem;
+  
+  &[data-theme="light"] {
+    background-color: $color-neutral-100;
+  }
+  &[data-theme="dark"] {
+    background-color: $color-neutral-950;
+  }
+
+  > .message {
+    @include flex-row-start-start;
+    width: 100%;
+    padding: .5rem;
+    gap: .5rem;
+
+    > .user-icon {
+      width: 2rem;
+      height: 2rem;
+      object-fit: cover;
+      border-radius: 50%;
+    }
+
+    > .message-infos-wrapper {
+      @include flex-column-start-start;
+      width: 100%;
+
+      &[data-selfie-message] {
+        align-items: flex-end;
+
+        .message-text {
+          background-color: $color-pink;
+          color: $color-white;
+          border-radius: .4rem;
+          border-top-right-radius: 0;
+
+          text-align: right;
+        }
+      }
+
+      > .first-row {
+        
+        > .username {
+          font-size: $font-size-normal;
+          font-weight: $font-weight-semi-bold;
+          &[data-theme="light"] {
+            color: $color-neutral-900;
+          }
+          &[data-theme="dark"] {
+            color: $color-white;
+          }
+        }
+        > .time {
+          margin-left: 1rem;
+          font-size: $font-size-small;
+          font-weight: $font-weight-normal;
+          &[data-theme="light"] {
+            color: $color-neutral-650;
+          }
+          &[data-theme="dark"] {
+            color: $color-neutral-600;
+          }
+        }
+      }
+    }
+
+    .message-text {
+      padding: .6rem;
+      background-color: $color-neutral-50;
+      border-radius: .4rem;
+      border-top-left-radius: 0;
+      box-shadow: rgba(17, 17, 26, 0.05) 0px 1px 0px, rgba(17, 17, 26, 0.1) 0px 0px 8px;
+
+      color: $color-neutral-900;
+      font-size: $font-size-small;
+      font-weight: $font-weight-normal;
+      
+      width: fit-content;
+      text-align: left;
+      word-break: break-word;
+    }
+  }
+}
